+++ services/oprf-dev-client/src/bin/world-id-dev-client-rp.rs
@@ -156,16 +156,22 @@ impl DevClient for WorldIdRpDevClient {
             .as_secs();
 
         let (uniquness_nullifier, session_nullifier) = tokio::join!(
-            self.components.authenticator.generate_nullifier(
-                &proof_request_uniqueness,
-                now,
-                Some(account_inclusion_proof.clone())
-            ),
-            self.components.authenticator.generate_nullifier(
-                &proof_request_session,
-                now,
-                Some(account_inclusion_proof),
-            )
+            self.components
+                .authenticator
+                .generate_nullifier_with_authorization(
+                    &proof_request_uniqueness,
+                    Some(account_inclusion_proof.clone()),
+                    setup.authorization_proof.clone(),
+                    now,
+                ),
+            self.components
+                .authenticator
+                .generate_nullifier_with_authorization(
+                    &proof_request_session,
+                    Some(account_inclusion_proof),
+                    setup.authorization_proof.clone(),
+                    now,
+                )
         );
 
         let uniqueness_epoch = uniquness_nullifier